Fixes language routing
