+++ src/multi-lang/ChangeLangLink.tsx
@@ -9,12 +9,15 @@
 import { useLocation } from 'react-router';
 import React, { useContext } from 'react';
 import { PreLangBasenameContext } from './PreLangBasenameContext';
+import { Link } from 'react-router-dom';
 
 function addLangToCurrentLocation(location, preLangBasename: string, lang: string) {
   const langRegexp = /\/(en|ko|jp)\//;
-  return location.pathname.match(langRegexp)
+  const locationWithProperLang = location.pathname.match(langRegexp)
     ? location.pathname.replace(langRegexp, `/${lang}/`)
     : `${preLangBasename}/${lang}${location.pathname}`;
+
+  return locationWithProperLang;
 }
 
 interface IProps {
@@ -25,5 +28,12 @@ export const ChangeLangLink: React.FC<IProps> = ({ lang, children, ...others }) 
   const location = useLocation();
   const preLangBasename = useContext(PreLangBasenameContext);
   // TODO : FUTURE : O.L : Change this 'a' to 'Link'
-  return <a href={addLangToCurrentLocation(location, preLangBasename, lang)} {...others}>{children}</a>;
+  return (
+    // <a href={addLangToCurrentLocation(location, preLangBasename, lang)} {...others}>
+    //   {children}
+    // </a>
+    <Link to={addLangToCurrentLocation(location, preLangBasename, lang)} {...others}>
+      {children}
+    </Link>
+  );
 };
